@@ -10,10 +10,10 @@ authors:
 description: A high-performance async/non-blocking Redis client for Dart.
 homepage: https://github.com/dartist/redis_client
 environment:
-  sdk: '>=0.8.10+6 <2.0.0'
+  sdk: ">=2.16.2 <3.0.0"
 dependencies:
-  logging: '>=0.9.0 <0.12.0'
+  logging: ^1.0.2
 dev_dependencies:
-  test: '>=0.12.0 <0.13.0'
-  mockito: '>=0.11.0 <0.12.0'
+  test: ^1.21.1
+  mockito: ^5.1.0
   benchmark_harness: 'any'
