@@ -20,7 +20,7 @@ class _CommandUtils {
     int keyValueStartIndex = firstParams != null ? firstParams.length : 0;
 
     int keysAndValuesLength = keys.length * 2 + keyValueStartIndex;
-    List<List<int>> keysAndValues = new List<List<int>>();
+    List<List<int>> keysAndValues = <List<int>>[];
 
     for (int i = 0; i < keyValueStartIndex; i++){
       keysAndValues.add(firstParams[i]);
@@ -37,12 +37,13 @@ class _CommandUtils {
 
   /// Returns a single list with the command and the arguments as bytes merged into one.
   static List<List<int>> mergeCommandWithStringArgs(List<int> cmd, List<String> args) =>
-    mergeCommandWithArgs(cmd, args.map((x) => UTF8.encode(x)).toList(growable: false));
+    mergeCommandWithArgs(cmd, args.map((x) => utf8.encode(x)).toList(growable: false));
 
   /// Returns a single list with the command, the key and the arguments as bytes merged into one.
   static List<List<int>> mergeCommandWithKeyAndStringArgs(List<int> cmd, String key, List<String> args){
     args.insert(0, key);
-    return mergeCommandWithArgs(cmd, args.map((x) => UTF8.encode(x)).toList(growable: false));
+    
+    return mergeCommandWithArgs(cmd, args.map((x) => utf8.encode(x)).toList(growable: false));
   }
 
   static List<List<int>> mergeCommandWithKeyAndArgs(List<int> cmd, String key, List<List<int>> args){
@@ -52,7 +53,7 @@ class _CommandUtils {
 
 
   static List<List<int>> mergeCommandWithArgs(List<int> cmd, List<List<int>> args){
-    List<List<int>> mergedBytes = new List<List<int>>(args.length + 1);
+    List<List<int>> mergedBytes =  [];
     mergedBytes[0] = cmd;
     for (var i = 0; i < args.length; i++){
       mergedBytes[i + 1] = args[i];
